pyproject.toml
[build-system]
build-backend = "poetry-core.masonry.api"
requires = ["poetry-core"]

[tool.coverage.report]
exclude_also = [
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "except Exception as e"
]

[tool.coverage.run]
omit = [
  "Final2x_core/__main__.py"
]

[tool.mypy]
disable_error_code = "attr-defined"
disallow_any_generics = false
disallow_subclassing_any = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
strict = true
warn_return_any = false

[tool.poetry]
authors = ["Tohrusky"]
classifiers = [
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
description = "core library for Final2x"
homepage = "https://github.com/Tohrusky/Final2x-core"
license = "BSD-3-Clause"
name = "Final2x_core"
readme = "README.md"
repository = "https://github.com/Tohrusky/Final2x-core"
version = "3.0.0"

# Requirements
[tool.poetry.dependencies]
PyYAML = "*"
ccrestoration = "0.1.1"
loguru = "*"
pydantic = "*"
python = "^3.9"

[tool.poetry.group.dev.dependencies]
numpy = "*"
pre-commit = "^3.7.0"
pyinstaller = "^6.3.0"
scikit-image = "*"
torch = "*"
torchvision = "*"

[tool.poetry.group.test.dependencies]
coverage = "^7.2.0"
pytest = "^8.0"
pytest-cov = "^4.0"

[tool.poetry.group.typing.dependencies]
mypy = "^1.8.0"
ruff = "^0.3.7"

[tool.poetry.scripts]
Final2x-core = 'Final2x_core.__main__:main'

[tool.ruff]
extend-ignore = ["B018", "B019", "RUF001", "PGH003", "PGH004", "RUF003", "E402", "RUF002", "B904"]
extend-select = [
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "PGH",  # pygrep-hooks
  "RUF",  # ruff
  "W",  # pycodestyle
  "YTT"  # flake8-2020
]
fixable = ["ALL"]
line-length = 120

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.isort]
combine-as-imports = true

[tool.ruff.mccabe]
max-complexity = 10
